// src/pages/Clients/Payments/DTOPayment.ts
export interface DTOPayment {
  no_invoice: string;
  period_start: string;
  period_end: string;
  amount: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  client_cedula: string;
  client_first_name: string;
  client_second_name: string;
  client_first_lastname: string;
  client_second_lastname: string;
  client_phone: string;
  client_email: string;
  admin_nit: string;
}